Joined all decorator classes in one file
