--- base_decorator.rb
@@ -1,7 +1,0 @@
-require_relative './nameable'
-
-class Decorator < Nameable
-  def initialize(namable)
-    @namable = namable
-  end
-end
